+++ imaging/CSVReader.h
@@ -1,6 +1,7 @@
 /*
 CSVReader.h
-Revised 8/3/2023
+Revised 9/14/2023
+Colton Cox - ccox60@uco.edu
 Nathan Wiley - nwiley@uco.edu
 */
 
@@ -22,6 +23,10 @@ void readIntArray(const char *fileName, int channels[NUMBER_OF_CHANNELS][2]);
 void readDoubleArray(const char *fileName, double doubleArray[NUMBER_OF_FREQUENCIES]);
 void readFloatArray(const char *fileName, float floatArray[NUMBER_OF_ANTENNAS][3]);
 void readComplexArray(const char *fileName, long double complex **complexArray, int *numRows, int numComplexNumbers[]);
-void readDataFiles(int channelsArray[NUMBER_OF_CHANNELS][2], double frequenciesArray[NUMBER_OF_FREQUENCIES] , float antennaLocationsArray[NUMBER_OF_ANTENNAS][3], long double complex **iqArray1, long double complex **iqArray2);
+void readDataFiles(int channelsArray[NUMBER_OF_CHANNELS][2], 
+                double frequenciesArray[NUMBER_OF_FREQUENCIES] , 
+                float antennaLocationsArray[NUMBER_OF_ANTENNAS][3], 
+                long double complex **iqArray1, 
+                long double complex **iqArray2);
 
 #endif