+++ changelog.md
@@ -19,5 +19,6 @@ and this project adheres to *[Semantic Versioning](https://semver.org/spec/v2.0.
 ### Removed
 
 ### Fixed
+- Imaging process
 
 ### Security